+++ limesium-reactive-logging/src/main/kotlin/eu/inqudium/limesium/reactive/logging/ExchangeLifecycle.kt
@@ -6,6 +6,7 @@ import eu.inqudium.limesium.common.EndpointLoggingMetrics
 import eu.inqudium.limesium.common.HeaderValueMasker
 import eu.inqudium.limesium.common.MdcKeys
 import eu.inqudium.limesium.common.NanoTimeSource
+import eu.inqudium.limesium.common.RequestLoggingProperties
 import eu.inqudium.limesium.common.Traceparent
 import eu.inqudium.limesium.common.failOpen
 import eu.inqudium.limesium.common.reportFailOpen