+++ limesium-reactive-logging/src/main/kotlin/eu/inqudium/limesium/reactive/logging/RequestLoggingWebFilter.kt
@@ -5,6 +5,7 @@ import eu.inqudium.limesium.common.HeaderValueMasker
 import eu.inqudium.limesium.common.MdcKeys
 import eu.inqudium.limesium.common.MdcScope
 import eu.inqudium.limesium.common.NanoTimeSource
+import eu.inqudium.limesium.common.RequestLoggingProperties
 import eu.inqudium.limesium.common.TraceMdcKeys
 import eu.inqudium.limesium.common.Traceparent
 import io.micrometer.core.instrument.MeterRegistry