+++ limesium-reactive-logging/src/test/kotlin/eu/inqudium/limesium/reactive/logging/CoRequestLoggingWebFilterTest.kt
@@ -10,6 +10,7 @@ import eu.inqudium.limesium.common.EndpointLoggingMetrics
 import eu.inqudium.limesium.common.HeaderLogProperties
 import eu.inqudium.limesium.common.MdcKeys
 import eu.inqudium.limesium.common.NanoTimeSource
+import eu.inqudium.limesium.common.RequestLoggingProperties
 import eu.inqudium.limesium.common.installMdcAdapter
 import eu.inqudium.limesium.common.keyValues
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry