+++ limesium-servlet-logging/src/test/kotlin/eu/inqudium/limesium/servlet/logging/RequestLoggingFilterAsyncTest.kt
@@ -10,6 +10,7 @@ import ch.qos.logback.core.spi.FilterReply
 import eu.inqudium.limesium.common.CapturedLogger
 import eu.inqudium.limesium.common.EndpointLoggingMetrics
 import eu.inqudium.limesium.common.MdcKeys
+import eu.inqudium.limesium.common.RequestLoggingProperties
 import eu.inqudium.limesium.common.keyValues
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry
 import jakarta.servlet.AsyncEvent